+++ db/migrate/20221012123642_add_close_date_to_project.rb
@@ -1,0 +1,5 @@
+class AddCloseDateToProject < ActiveRecord::Migration[7.0]
+  def change
+    add_column :projects, :closed_at, :datetime
+  end
+end